classifier_app/requirements.txt
altair==4.1.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
attrs==19.3.0
backcall==0.2.0
base58==2.0.1
bleach==3.3.0
blinker==1.4
boto3==1.14.30
botocore==1.17.30
cachetools==4.1.1
certifi==2020.6.20
cffi==1.15.1
chardet==3.0.4
click==7.1.2
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
docutils==0.15.2
entrypoints==0.3
enum-compat==0.0.3
filelock==3.0.12
future==0.18.2
idna==2.10
importlib-metadata==5.1.0
ipykernel==5.3.4
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.5.1
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
MarkupSafe==1.1.1
mistune==0.8.4
munch==2.5.0
nbconvert==5.6.1
nbformat==5.0.7
nest-asyncio==1.5.6
notebook==6.4.12
numpy==1.21.0
packaging==20.4
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.1
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.15.0
ptyprocess==0.7.0
pyarrow==1.0.0
pycparser==2.21
pydeck==0.4.0
Pygments==2.7.4
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.4
pyzmq==19.0.1
requests==2.24.0
s3transfer==0.3.3
Send2Trash==1.8.0
six==1.15.0
streamlit==0.64.0
terminado==0.8.3
testpath==0.4.4
toml==0.10.1
toolz==0.10.0
torch==1.6.0
torchvision==0.7.0
tornado==6.0.4
traitlets==4.3.3
typing-extensions==4.4.0
tzlocal==2.1
urllib3==1.25.11
validators==0.16.0
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
zipp==3.11.0
